Update TrieTest to test iterative findNode
We're just about to remove the recursive implementation, lets make
sure we're testing what the app is going to use.

Note: this commit changes the simpler (non-domain) tests to reverse
Strings before insertion into the Trie. Trie is now completely domain
specific, hence that is warranted. (In a later step, put() will be
updated to accept plain Strings, it will also take care of reversal
as appropriate.)