Move Trie search into TrieIterator
For the blocklist, we only need to check if any node in the Trie
matches the input domain (multiple nodes can match if e.g. the root domain
and a subdomain are stored in the Trie). However the whitelist can
have multiple layers, e.g. root domain whitelists one or two domains,
a subdomain can whitelist additional domains - that requires
finding all nodes matching a given input, which TrieIterator implements.
(EntityList will be upgraded to use this iterator in a later commit, it
 currently has its own Trie search implementation - this iterator means
 we only need one implementation for normal and whitelist Trie search.)